task 4.21

+++ Osa4/blogilista/controllers/blogs.js
@@ -44,9 +44,23 @@ blogsRouter.post('/', async(request, response,) => {
   response.json(savedBlog)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+blogsRouter.delete('/:id', async (request, response, next) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  const user = await User.findById(decodedToken.id)
+
+  const blogToDelete = await Blog.findById(request.params.id)
+
+  if ( blogToDelete.user._id.toString() === user._id.toString() ) {
+    try {
+      await Blog.findByIdAndRemove(request.params.id)
+      response.status(204).end()
+    } catch (exception) {
+      next(exception)
+    }
+  } else {
+    return response.status(401).json({ error: 'Unauthorized' })
+  }
 })
 
 blogsRouter.put('/:id', async (request, response, next) => {
